sim: machine_context

+++ src/sim/platform.cpp
@@ -12,7 +12,7 @@ namespace sl::exec::sim {
 
 meta::result<platform, std::error_code> platform::make() {
     errno = 0;
-    const int page_size_result = ::sysconf(_SC_PAGESIZE); // never indeterminate
+    const long page_size_result = ::sysconf(_SC_PAGESIZE); // never indeterminate
     if (page_size_result < 0) [[unlikely]] {
         return meta::errno_err();
     }
